bug fixes, input handler

+++ src/inventory.h
@@ -67,7 +67,7 @@ SISTERRAY_API void init_item_type_data();
 SISTERRAY_API i16 sort_inventory(i32 sort_type);
 bool usable_in_inventory_menu(u16 item_id);
 u16 get_restriction_mask(u16 item_id);
-u16 get_target_flags(u16 item_id);
+u8 get_target_flags(u16 item_id);
 void updateInventoryEntry(u16 item_id, u8 quantity);
 
 #endif
